+++ geotrek/sensitivity/tests/test_functional.py
@@ -29,11 +29,13 @@ class SensitiveAreaViewsTests(GeotrekAPITestCase, CommonTest):
         }
 
     def get_expected_json_attrs(self):
+
         return {
             'attachments': [],
             'contact': '<a href="mailto:toto@tata.com">toto@tata.com</a>',
             'description': 'Blabla',
             'kml_url': '/api/en/sensitiveareas/{}.kml'.format(self.obj.pk),
+            'openair_url': '/api/en/sensitiveareas/{}/openair'.format(self.obj.pk),
             'publication_date': '2020-03-17',
             'published': True,
             'rules': [
@@ -127,6 +129,7 @@ class RegulatorySensitiveAreaViewsTests(GeotrekAPITestCase, CommonTest):
             'contact': '<a href="mailto:toto@tata.com">toto@tata.com</a>',
             'description': 'Blabla',
             'kml_url': '/api/en/sensitiveareas/{}.kml'.format(self.obj.pk),
+            'openair_url': '/api/en/sensitiveareas/{}/openair'.format(self.obj.pk),
             'publication_date': '2020-03-17',
             'published': True,
             'rules': [