+++ geotrek/sensitivity/tests/test_parsers.py
@@ -72,6 +72,7 @@ json_test_species = {
             "structure": "LPO",
             "species_id": 7,
             "kml_url": "https://biodiv-sports.fr/api/fr/sensitiveareas/46.kml",
+            "openair_url": "https://biodiv-sports.fr/api/fr/sensitiveareas/46/openair",
             "geometry": {
                 "type": "Polygon",
                 "coordinates": [[[3.1, 45], [3.2, 45], [3.2, 46], [3.1, 46], [3.1, 45]],
@@ -102,6 +103,7 @@ json_test_species_page_2 = {
             "structure": "LPO",
             "species_id": 7,
             "kml_url": "https://biodiv-sports.fr/api/fr/sensitiveareas/47.kml",
+            "openair_url": "https://biodiv-sports.fr/api/fr/sensitiveareas/47/openair",
             "geometry": {
                 "type": "MultiPolygon",
                 "coordinates": [[[[3.1, 45], [3.2, 45], [3.2, 46], [3.1, 46], [3.1, 45]],